split evidence from misbehaviour
