+++ x/ibc/02-client/types/tendermint/doc.go
@@ -1,5 +1,5 @@
 /*
-Package tendermint implements a concrete `ConsensusState`, `Header` and `Equivocation`
-for the Tendermint consensus algorithm.
+Package tendermint implements a concrete `ConsensusState`, `Header`,
+`Misbehaviour` and `Equivocation` types for the Tendermint consensus light client.
 */
 package tendermint
